refactor: split up advisory and align with new product impl
